[issues/54] Refactoring for testing
